% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_data_fun.R
\name{construct_data}
\alias{construct_data}
\alias{transform_factor}
\title{Data construction functions}
\usage{
construct_data(formula, data)

transform_factor(formula, data)
}
\arguments{
\item{formula}{The formula of the model}

\item{data}{The original dataset}
}
\description{
These functions are used to construct the data from a given formula. Handles change from factor into several dummies for example.
They also create reference individuals at which to evaluate the MTE and MTR (if no default is provided).
For the numerical variables, take the average of the variable; for the factors, take the first level.
}
