% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_transform_fun.R
\name{Kappa_fun}
\alias{Kappa_fun}
\alias{kdu_transform_fun}
\alias{Kappa_homogenous_fun}
\alias{ku_transform_homogenous_fun}
\title{Control functions transformations for selection probabilities}
\usage{
Kappa_fun(p, pol_degree = 5)

kdu_transform_fun(u, d, pol_degree = 5)

Kappa_homogenous_fun(p, pol_degree = 5)

ku_transform_homogenous_fun(u, pol_degree = 5)
}
\arguments{
\item{p}{Vector of propensity scores to transform into a flexible function.}

\item{pol_degree}{Degree of the polynomial transformation.}

\item{d}{Which potential outcome to consider (only needed for \eqn{k_d(u)} with heterogenous treatment effects).}
}
\description{
These functions provides pre-specified transformations to control flexibly for the selection probabilities in the regression. \cr
These correspond to \eqn{\kappa_d(p)} and corresponding \eqn{k_d(u)} in Bruneel-Zupanc (2024). \cr
Special functions are used for homogenous treatment effect specifications because the code is different.
July 2024: for now, only polynomial transformations are encoded.
}
\details{
See Andresen (2018) or Bruneel-Zupanc (2024) for computation details linking \eqn{\kappa_d(p)} and the corresponding corresponding \eqn{k_d(u)}. \cr
\eqn{\kappa_1(p) = E(U_1 | U_D \leq p) } and \eqn{\kappa_0(p) = E(U_0 | U_D>p) } while \eqn{k_d(u) = E(U_d | U_D=u)}. \cr

In the case of homogenous treatment effects: \eqn{k_0(u) = k_1(u)}. This provides some restriction on \eqn{\kappa}, hence the special functions.
}
\examples{
v = seq(0.1, 0.9, by=0.1)
# Transformations for general Heterogenous TE functions:
Kappa_fun(p=v, pol_degree=6)
k1u = kdu_transform_fun(v, d=1, pol_degree=6)

# Transformations for Homogenous TE functions:
Kappa_homogenous_fun(p=v, pol_degree=6)
ku = ku_transform_homogenous_fun(v, pol_degree=6) # no d anymore, same for both d here;
}
