% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mte_plot.R
\name{mtr_plot_fun}
\alias{mtr_plot_fun}
\alias{mte_plot_fun}
\title{Marginal Treatment Effect (MTE) and Responses (MTR) plots}
\usage{
mtr_plot_fun(dat_plot, common_supp)

mte_plot_fun(dat_plot, common_supp)
}
\arguments{
\item{dat_plot}{Data frame with the estimated MTE and MTR values and their confidence intervals. Must contain specific variables: \code{Phat}, \code{mtr0}, \code{mtr1}, \code{mtr0_lwr}, \code{mtr1_lwr}, \code{mtr0_upr}, \code{mtr1_upr}, \code{mte}, \code{mte_lwr}, \code{mte_upr}.}

\item{common_supp}{Vector of two values indicating the common support of the plot. Default is the full support \link{0,1}.}

\item{conf_band}{Indicates whether to plot the confidence intervals. Default is "TRUE".}

\item{colMTE, colD0, colD1}{Color of the MTE, MTR0 and MTR1 curves.}
}
\description{
Plot the MTR and MTE estimated curves with their confidence intervals.
}
\details{
Attention: by default in \code{semiivreg} the confidence intervals are computed analytically, and include an error because the first stage propensity score.
This is corrected in \code{semiivreg_boot} by bootstrapping the entire estimation to obtain the confidence intervals.
}
